AUT-212: ECS rollout configuration tweaks in build and integration only
Currently waiting 300s for the apps to drain during a rollout as the deregistration delay is set to the default.  Frontend transactions complete in a matter of seconds so there is no reason to wait for 5 mins.  This will reduce rollout timeout.

Frontend apps start quickly but a small grace period makes sense just in case  to allow for connections.