GUA-574: Update account blurb on sign-in-or-create
The accounts team are going live with some updates to the account
management area including a content update on the account blurb which
now lists the services the account can currently be used with.

Implement the same update on the authentication frontend side so that
the entire journey is congruent.