AUT-1720: don't try to export invalid tfvars
Add a check to `read_secrets.sh` to ensure that the secret name is a
valid environment variable name (eg. a not secret name with a hyphen in
it) before attempting to export the variable.

If the name is not a valid `bash` variable name, the script errors and
exits.

Previously, the script would error and exit anyway, but the error was
not particularly helpful (`'some-variable=something': not a valid
identifier`).