AUT-1715: Align Sign in or create page content for both MANDATORY and OPTIONAL clients
OPTIONAL clients (where the user is not required to create an account before starting the service journey) had slightly different content on the sign in or create screen, to cover the scenario where the user was to create an account to save their progress.  As this content was not thought to be help it has been removed, so that OPTIONAL clients now see exactly the same content as MANDATORY clients.