AUT-1653: Change logging approach to debug Build
Since the integration PR has been merged we've found that form submissions
on Build are resulting in a "Sorry, there is a problem" being presented to
users.

Looking into this, it seems there's a call to undefined properties
of `error`, so I've changes the syntax to optional chaining.

The hypothesis is that this is a result of the shape of values
that are being provided via Secrets Manager so I'm also logging the values
of those.