AUT-1816: Allow SmartAgent to be used based on clientId
At a meeting on 6 October we were asked if it would be possible for form
submissions to SmartAgent to be tested in the Live environment before
go-live. This requirement had not been known before the meeting and
introduces some complexity because the current implementation relies
upon a switch at the environment level: an environment will either send
submissions to SmartAgent or Zendesk, not both.

Having discussed this with the Tech Lead it initially seemed best to introduce
a controller method that would route to SmartAgent where the clientId matched
that of the production stub. However, trying to implement this was problematic
and I couldn't get it to work.

I then thought of additional options and concluded the easiest way to achieve
this would be to set the form action attribute to a new route where the condition
was met, and to use the existing contactUsQuestionsFormPostToSmartAgent method
as the handler for that route.