AUT-1650: Fix back links from Triage Page
When testing inbound links from the Triage Page in Integration today
we found that links which include the `theme=id_check_app` always use
`/contact-us` in the back link.

Looking into this, it seems result from this value being hard-coded into
the template.

This commit sets the back link to use the referer where it is a valid
referer or, where it is not present or valid, to use `/contact-us`.