AUT-1338: Update text on problem creating a login radio button
Updates text from "You do not have a UK mobile phone number" to "You had another problem with a phone number".

Users can now create a GOV.UK One Login with a non-UK mobile number, so the lack of a UK number should not, in itself, be a problem. However, users are still selecting this option, so users could still be experiencing problems that are not covered by the other radio buttons (the verbatim data from the user support form does not give enough detail to know what the problems are). Keeping the radio button but changing the wording may help us understand some of these other problems are.

This also changes the order of the radio buttons on the account creation support page so that the new label makes sense in the context of the other labels