AUT-2113: Redirect requests to login when reauthentication is requested
Updates:
* state machine
* authorize controller

Temporarily route to sign-in-or-create pending decision on whether to include the new reauthentication screen.

Note that this behaviour is technically behind a feature flag. However, since currently we have no separate page for the re-auth flow (both re-auth and normal auth go to /sign-in-or-create for now, there is no impact since we will be setting authenticated to false for reauthentication journeys in the backend.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>