AUT-2142: Change error screen in specific scenario (Detail below)
This commit changes the path users are routed to when the POST of
their password results in the ENTERED_INVALID_MFA_MAX_TIMES error
code. It does not change the path users are routed to when this
error code is encountered in other scenarios.

Because we now need to show different screens based on where the user
encountered the error, we no longer have a 1:1 mapping between error
codes and paths.

I have opted to introduce a new `getErrorPathByCodeForPasswordFailures`
proxy function here rather than adding logic to `getErrorPathByCode`,
because the latter approach would require an additional argument and
clutter