AUT-1553: Extend getErrorPathByCode to accept an override
The need to show different error screens in response to a single
error code breaks the 1:1 relationship between codes and paths that
getErrorPathByCode relies upon.

This commit adds an optional second parameter that allows a specific
error code to be overridden should it be encountered in the context of
the call.