BAU: Remove the deprecated 'landing' component
START and the landing component are now deprecated post the auth orch split.

Users are manually navigating to the url 'signin.account.gov.uk' after they have started their journey.  Before this would have triggered another journey start, letting the users login correctly, but since the split this no longer works as additional work is done in the new /authorize component that is missing in 'landing'.  As a result users will then see an error if trying to login.

This PR changes the behaviour to redirect to the start of the journey (sign-in-or-create), without making any changes to the session or calling the start api that used to happen.  All these things have now moved to the /authorize endpoint instead.  As any existing session is not overwritten users can start the journey again and complete it without error.  Redundant code in the 'landing' component has been removed.

If anyone tries to access 'signin.account.gov.uk' directly, without having come from a service via orchestration authorize, they will now see an error page as this is not a valid journey.