AUT-2370: Wire account-intervention middleware into prove identity route
Currently, the account interventions middleware is called in /auth-code. However, in cases where a user needs to re-prove their identity, this journey happens before /auth-code is hit, and we should circumvent flagged users from continuing their journey at this point.

This also corrects a small issue in the middleware where the redirects were not being returned by the function, which caused issues in tests.

Co-authored-by: Matt Votsikas Mclean <matt.votsikas-mclean@digital.cabinet-office.gov.uk>