BAU: Rework local stub to call /authorize
The local stub still relies on starting the journey from the root '/' which redirects to 'sign-in-or-create', a journey that is soon to be removed.

Instead make the local stub start with a call to /authorize just like orchestration.

The local stub is at the end of its useful life and needs to be replace with a real client, but this change will keep it going for a bit more.  After login there will still be an error as there is no orchestration stub or real client to redirect back to, but the journey successfully completes /auth-code and hands back to orchestration.

As a key is required to decrypt the authorize jwt the local application needs a connection to the aws account to read the key, so the startup script needs to retrieve credentials for the account in order to work.

This changes fixes local startup but breaks startup of the frontend in docker-compose, which still needs fixing.