BAU: make dev docker exposed ports modifiable
Previously you had to modifiy the docker-compose.yml file to change the
exposed ports for the frontend services. This change makes the ports
configurable via the .env file.

Rather than having the environment variables in the 'list' format,
instead define them as a key value pair. This is to make it easier to
read and understand the environment variables that are being set, due
to syntax highlighting.

NB: json values must be surrounded by single quotes, otherwise the
compose file will not be valid.

Add checks for required environment variables in the docker-compose.yml
file. This is to ensure that the application will not start if the
required environment variables are not set. It also signposts the user
to where they should be set / retrieved from.

Add a test for `.env` file to ensure that the 'ask for this value'
variables have been set. Additionally, test that
$ORCH_TO_AUTH_SIGNING_KEY is defined as a multiline string, not with
`\n` literals.