AUT-2552: Add applied at timestamp to interventions response
This updates the response we're expecting from the account interventions service in the frontend api to add the timestamp that the intervention was applied at.

This field is currently unused, but will be used as part of the logic to ensure that a user who has just reset their password is not made to go through the reset password journey twice.