Add defaults to the port variables in startup.sh
Additionally, ensure `startup.sh` and `shutdown.sh` are using
`set -euo pipefail` to catch errors and unset variables.

Finally, rather than pre-setting the `ACTION_*` variables, we can
set defaults in the `if` statements. This means that when updating /
reading the script, we can see the default values at the point of use.