AUT-2549: Added ipv callback journey from auth code
Added PATH_NAMES.PROVE_IDENTITY as a potential target path following on
from RESET_PASSWORD given that isIdentityRequired == true.
This change fixes the issue where the invalid journey was causing the final screen of the
reprove identity journey to loop. Now a user will be allowed to access
the service and complete their journey.