AUT-2623: Account for reproveIdentity in verify code controller AIS block
We check for the suspensions status of a user when they reset their password - if they are suspended, they should not be able to change their MFA method. However, if they are suspended because they need to take a user action (for example, they are required to reset their password or reprove their identity), we should allow them to change their MFA method so that they can go on to do this user action.

We were previously accounting for this by ignoring a suspensions status that came with a reset password intervention - this change also adds in logic to ignore a suspensions status in the verify code controller that comes with a reprove identity intervention. This will mean that a user who is required to reprove their identity will be allowed to change their MFA method before going on to prove their identity.