AUT-2598: Update success status code for reset password check email service
Currently, the api response for a reset password request returns a 204 (no content). We want to start returning some content, so we'll need to support it returning a 200. This adds 200 as a possible API response code, as a preparatory step before we start returning a 200 from the API.

When we have switched the API response over, we can remove 204 as a possible response code from here.