AUT-2598: Set mfa details in session in reset password check email controller
Currently, if a user:
* locks themselves out of their account due to too many passwords and then;
* attempts to access the service again in a new session and;
* chooses to reset their password when they hit the blocked screen

They will not be prompted to enter their 2fa during the password reset journey. This is because we determine whether someone needs to enter 2fa depending on whether the enterEmailMfaType is set in the session. This is normally set by the response from the "CheckUserExists" api endpoint, but when someone is locked out, this response returns an error and we do not get the mfa detail.

This commit forms part of a fix for this. We've updated the resetPasswordRequest response to contain the user's mfa details. Therefore, in this commit we:

* Add a new response interface for this request to reflect the data coming back in the body
* When we don't already have the mfa information set in the session, we set it here from the data contained in the body.

This will mean that, when a user enters the code from their email, the state machine will return the enter mfa code pages next, rather than going straight to the page where a user resets their password.