BAU: Remove international phone numbers middleware
This was used to set a variable that controlled feature-flagging behaviour in the views. We no longer need this since we have removed the logic which acts on this feature flag as the support for this has been on in all environments for a while, so we are safe to remove the feature flag.