AUT-2602: create session locks based on the lockout information returned by CheckUserExists API
The CheckUserExists API is returning information about the AuthApp locks at the moment, for the SIGN_IN and RESET_PASSWORD journeys. I have named the variable in the UserExists response to authAppLockoutInformation, the name of this should be changed if other lockout information is added to this API. Did this atm to make the code easier to understand.