AUT-1984: Add debug logging to sidecar tests
If the `RUNNER_DEBUG` environment variable is set to `1`, the logs from
the sidecar container will be streamed to the console. This is useful
for debugging test failures.

This is automatically set when the GitHub Actions job is rerun with the
"Enable debug logging" option checked. Otherwise, the test can be run
on the CLI with `RUNNER_DEBUG=1 yarn test:sidecar`