AUT-2686: Explicitly call save on a session in state machine
Sometimes when going through a user journey, a user can click continue but end up with an invalid transition error in the allow-user-journey-middleware.

We think this is because the call to session.save (which is async and happens by default at the end of every express http request) has not yet happened, and so when we check where we think a user should be from the session (which is out of sync), the application returns an invalid journey error, despite the user being on a valid journey.

This commits aims to stop this race condition, by explicitly calling session.save once we've updated the user's journey. This is referenced in the documentation[1] (see Session.save) on redirects. This necessitated making some functions async, and relevant changes in tests.

[1]: https://expressjs.com/en/resources/middleware/session.html