AUT-2733: Set consent cookie HttpOnly attribute to false
The HttpOnly attribute of the `cookie_preferences_set` cookie
was set to true in commit c2fbdc under AUT-1985 to address concerns
from an IT Health Check linked to in AUT-1949.

The IT Health Check stated:

"By acquiring another user's session cookie value, an attacker
 may be able to connect to the application as that user." and "Cookies
 within the application were found to lack the Secure and HttpOnly
 flags, leaving it at risk of compromise. The cookies were
 identified as not being responsible for controlling user sessions
 for the application"

Subsequent testing has revealed the cookie banner does not
function as expected where the `cookie_preferences_set` is
sent by the server with a HttpOnly attribute. This is because the
cookie banner functionality relies on JavaScript to update the cookie
in response to user actions.