AUT-2686: Improve logging message for invalid journey
The way this was worded was misleading, indicating that the path in the user's session is the path they're coming from, and the request path is the one they're going to. This is confusing, as actually they should be the same (the session stores what path the state machine computes the user should now be on, and this is compared against the request). Reworded to clarify.