AUT-2772: Add addLanguageParam as global in nunjucksEnv
To achieve this it was necessary to use `require()` rather than the
`import` used elsewhere in the codebase because the dependency does
not currently support ESM.

Attempting this initially resulted in this error:

 "Error [ERR_REQUIRE_ESM]

  require() of ES Module not supported. Instead change
  the require of language-param-setter.js  to a dynamic import()
  which is available in all CommonJS modules."

After some investigation and reviewing the implementation used in the
Mobile Pod, I found that changing the casing of compilerOptions.module
in tsconfig.json to what is shown as the default in the TypeScript docs
allowed the import without error. This is despite the previous casing
being included in the 'Allowed' list.

- Mobile Pod implementation: https://github.com/govuk-one-login/mobile-id-check-frontend/pull/937/files
- compilerOptions.module in TypeScript docs: https://www.typescriptlang.org/tsconfig/#module