AUT-2578: Use in-built URL functionality to construct url
We need this to call the  in the common headers library. Using the url library to construct this avoids us having custom code to create the url (it automatically handles any combination of the base path and path having leading / trailing slashes). The library uses this URL to extract a domain, so technically the full url is not currently strictly necessary. However, we're correctly constructing this for additional robustness in case future updates to the library use this url for additional purposes.