BAU: Prevent extra arguments in startup.sh
At some point, it became common practice to pass extra arguments to
`startup.sh`.

This behaviour does not change the behavior of the script, but can lead
to confusion, as users may believe that this extra argument will cause
something to happen, and may be confused when it does not.