AUT-2578: Ensure helper function constructing http config respects base url when passed in
This ensures that when we call the createPersonalDataHeaders library function, we pass in a url constructed with the correct base url. In some cases we pass in a different base url, for example when we call the orchestration backend, so we want to ensure this is respected in all logic within this config function.

Currently this has no effect on the output of the createPersonalDataHeaders function, but it does affect logging within the library which logs which domains the headers are being forwarded to, so this ensures these logs are accurate.