BAU: Further improvements to error handling
This commit adds a new custom exception, `FatalError`, which is raised
when a fatal error occurs. This is used to catch all exceptions that
should cause the script to exit, rather than calling sys.exit directly.

Additionally, identify better when lack of VPN connection is the cause
of an error, and correctly report this to the user. There are still some
hand-wavey 'unexpected error' messages, but in these cases, the stack
trace is printed, which will hopefully better help diagnose the issue.

Finally, stop importing the specific `botocore.exceptions` exceptions
individually, as it was becoming unwieldy.