AUT-2578: Convert reset password service to use common headers
This means that we submit txma security headers in requests to the backend.

Note that the interface of the reset password service specifies an isForcedPassword param, which wasn't specified on the implementation. Making this change required us to add this parameter too otherwise our tests would not compile. However, the isForcedPasswordReset parameter is unused which is a small bug in our service which we'll fix subsequently