AUT-2797: Remove terms that trigger iOS bug from labels
There is a bug in Safari on iOS that causes users be be presented
with a prompt to 'Scan card' when some input fields receive focus.

Testing has revealed this impacts only fields containing the term
'security code'. It seems this is causing browser autcomplete heuristics
to interpret the field purpose as being for a CVV code.

This commit removes the term 'security code' from all form labels
and introduces tests to ensure it is not accidentally reintroduced.