AUT-2682: Replace positional locale keys in accessibility statement locales
Changes to the accessibility statement required for the launch of
webchat will require new sections to be added between existing sections.
This highlights an issue with using keys such as `section3`, `section4`
etc. because they'll need to be shifted to accommodate new / removed sections.

This PR replaces positional keys such as these with descriptive names that
do not reference position. I've limited these changes to the accessibility
statement section of locale files to keep the changes of this PR focused on
the Jira ticket and reduce the risk of breaking something on a page unrelated
to the PR (the design review will only look at the Accessibility Statement).