AUT-3401: Replace auto refresh with HTML form
This fixes the accessibility issue arising from automatic
page refreshes by introducing a form that will instead allow
the user to refresh the page when they are ready to do so.

For the time-being, this new form is behind a feature flag.

Rather than creating a new controller method, it seemed
sensible to repurpose the existing method to handle both
GET and POST requests.

This version that will be progressively enhanced with
JavaScript to improve the user experience.