AUT-1723: trust cloudfront cidrs in sidecar
As we now have a cloudfront distribution in front of the ALB, we need to
add the cloudfront node CIDRs to the list of trusted proxies in the
sidecar container. This will allow the sidecar to correctly identify the
client IP address via X-Forwarded-For headers.

With cloudfront, the sidecar sees the following headers:

X-Forwarded-For: client_ip, cloudfront_ip, eg:
X-Forwarded-For: 1.2.3.4, 15.158.48.119.

As 15.158.48.119 was previously not in the list of trusted proxies,
nginx treated it as the client IP address (as it is the rightmost IP
that's not a trusted proxy). This resulted in the sidecar requesting
basic auth. Now these IPs are marked correctly, it will work as expected
again.