AUT-2790: Add logic to redirect to sign-out when introducing wrong email in re-auth journey.
For a re-authentication journey the use must present the same email address that they used to sign in with.

If the user presents an incorrect email address too many times when on the re-authentication journey they will be logged out by redirecting to orchestration.

Also continued to make the integration tests more robust by ensuring the initial call to the app in the before method is waited for and the env variables that control the feature switches are explicitly set.