AUT-2789: Log user out when they enter their password incorrectly max allowed times.
In order to handle too many password entry failures by the user during a reauth flow but still maintain the lock out for failures on all other flows we will use the journey type to determine the journey the user if on.  If on the reauth journey we will redirect to orchestration with the new path parameter, error=login_required, that orchestration will indicate that the user should be logged off.

Also continued to make the integration tests more robust by ensuring the initial call to the app in the before method is waited for and the env variables that control the feature switches are explicitly set.