AUT-3046: Corrected behaviour for handling when sending an MFA code errors
Removed the return statement from the error handler in send-mfa-controller as the handler should only set the redirect path and allow the error to be passed back to express when the BadRequestException is thrown.

Moved the pageNotFoundHandler above the error handlers to make it clear that this is not an error handler.  Added a comment to reinforce the fact that there are only two error handlers.

Updated tests to reflect these changes.