AUT-3401: Move spinner container into form
While reviewing the code, the Orchestration Tech lead asked how the
JavaScript button would work in the completed state. This question
revealed an issue with the existing static HTML structure because the
button will rely on a containing form to perform the submit action.

This commit simply moves the spinner container into the form