AUT-3052: Fix disappearing links
Previously, when either sms mfa or an auth app code were submitted incorrectly, the link to reset mfa would disappear.

This is because when the template was first rendered (before an incorrect code was submitted), its value would be of type boolean. However, we were storing the cached version as a string, meaning that on page reload due to incorrect code, the equality check would no longer pass.

This fixes this by saving the value as a boolean, rather than a string.

Need to check why the equality condition needs to be changed