AUT-3052: Fix disappearing account recovery links on mfa
Previously, when either sms mfa or an auth app code were submitted incorrectly, the link to reset mfa would disappear.

This is because when the template was first rendered (before an incorrect code was submitted), its value is a boolean. However, when we store the cached version (which gets used when the page reloads e.g. for a validation error) the value is a string.

This meant that on page relaod, the equality check (=== true) would return false, and the account recovery content would disappear.

I tried storing the cached value as a boolean rather than a string (e.g. changing

`<input type="hidden" name="isAccountRecoveryPermitted" value="{{isAccountRecoveryPermitted}}"/>`

to

`<input type="hidden" name="isAccountRecoveryPermitted" value={{isAccountRecoveryPermitted}} />`

This still did not work, as under the hood nunjucks converts these cached values to strings anyway, so we still need to either convert the cached value to a boolean, or adjust the equality check to check for strings and booleans.