AUT-2914: Add middleware for handling missing or invalid csrf tokens
Currently, we return a server error in this case, as we do not have any specific handling for this case. When we get spam requests, this means we end up return 5xxs and flagging errors.

Adding middleware to handle this case means that we will return 403s, rather than 500s. This solution is in line with the suggestion from csurf: https://www.expressjs.com.cn/resources/middleware/csurf.html