BAU: Disable form-action CSP directive default
The upgrade to Helmet has introduced a number
of sensible CSP directive defaults including new
directives not previously present in authentication
frontend:

* base-uri 'self'
* font-src 'self' https: data:;
* form-action 'self';
* upgrade-insecure-requests

The full list of default directives can be seen on their
GitHub page:

https://helmetjs.github.io/#content-security-policy

After the PR to upgrade Helmet (#1986) was merged
we found that the “form-action 'self’” directive was
causing acceptance tests to fail in the CI pipeline with
this message: “Refused to send form data to
'https://signin.build.account.gov.uk/enter-password'
because it violates the following Content Security
Policy directive: "form-action 'self’”.”

The PR was therefore reverted in #1998.

This issue was not happening in manual testing and it is
not clear why the acceptance tests would fail: the form
action attribute uses a relative and unqualified path (so
must therefore be using the same origin as the document
being served, including the same URL scheme and port
number).

Because this directive was not present before the Helmet
upgrade and I cannot find why it would be causing the
acceptance tests to fail, it seems most pragmatic to remove
the directive - which is what this commit does.