AUT-3665: Add helper to transform templates for mobile
This helper is intended to provide a single mechanism that can be
added to controllers as required. The helper receives the web template
and a boolean reflecting context as arguments. It will returning the
path to the mobile equivalent template where the context is mobile.

The intention is for the helper to reinforce a consistent approach to
introducing mobile templates.