AUT-2141: Extract password guidance to shared component
While working on this ticket I noticed that the password guidance
details component is duplicated across multiple pages, with the
same code being replicated in templates and translation files.

In addition to making the content update (adding an additional
guidance paragraph), this commit creates a shared component and
removes the duplication that existed across the create and reset
password templates and translation blocks.