AUT-3725: Redirect to error page if enter password response returns missing or invalid session
This error response means that we cannot continue. However, at the moment we fall through to the default error behaviour, which is to reload the page with a validation error, making it seem like the user entered the incorrect password.

This handles an edge case in reauthentication whereby the user is logged in in two tabs, starts a reauth journey in another up until the point of entering a password, and then logs out in the first tab. At this point, they will enter an infinite loop of seeming password validation errors regardless of password, whereas in fact the backend is totally unable to proceed.

It's also possible to replicate this in the signin journey, although you have to be very specific with timings.