BAU: Prevent Sass mixed declaration warnings
The CSS Working Group has changed the way CSS handles declarations
mixed with nested rules. Sass has been updated to reflect this change
and any declarations that appear after nested rules are deprecated
in Dart Sass distributions > 1.77.7 (See [^1] below)

As a result of this we have been seeing multiple deprecation warnings
when compiling our Sass. This commit simply changes the code so that
a nested rule appears after declarations in the same block.

[^1] https://sass-lang.com/documentation/breaking-changes/mixed-decls/