BAU: Remove logic for non 2 hour lockout from index-wait template
This feature flag is on in all environments, and we no longer need to support this being false. This is part of removing this feature flag from our codebase.

This commit also removes the flag from being passed through in the options when rendering this template

Also removes the translations for the elements on this page which were used when the feature flag was off, which are no longer used in this application