AUT-3760: Rollout `supertest` wrapper to all uses
The previous commit introduces a wrapper for `supertest` that will append most tests with a snapshot check for their GA4 taxonomies.

Here we roll out this new wrapper to all uses of `supertest`.

This establises a baseline of taxonomy snapshot testing across all integration tests, giving us some ability to check for regressions when changing the logic for calculating taxonomies.

Perhaps we can remove or reduce this level of snapshot testing in time as we understand more and centralise the logic of computing taxonomies.