AUT-3760: Introduce `getRequestTaxonomyMiddleware`
A new middleware is introduced that will look at the request object provided and determine the GA4 taxonomy values based on the request. These are then provided to the choosen page template to use in the GA4 script.

This commit applies the middleware to just a few selected routes. Later commits will move the middleware application to the "global" app level, not individual routes.

A few simple routes are picked to start with, replacing the hardcoded default taxonomies of `"authentication"` & `""`.