AUT-3760: Apply `getRequestTaxonomyMiddleware` to all routes
All pages rendered now have taxonomies calculated by the `getRequestTaxonomyMiddleware`, which uses the `request` object.

It was necessary to do these remaining pages in one go, hence the large commit. Many use shared templates, validator middleware or redirects to shared error pages. This shared nature of this set of pages is the primary reason for this refactor, as we will be layering a new set of `reauth` taxonomies soon on several of these pages.

This isn't a clean refactor. Changes are made which correct gaps in existing taxonomies (due to the shared components which were not able to include taxonomies previously). It would be extra  work to get the middleware to replacate the existing taxonomy gaps.

Here's a list of the changes in addition to middleware application changes:
- Errors on the `resend-mfa-code`, `enter-email`, `enter-email-create-account`, `check-your-email`, `check-your-email-security-codes`, `resend-email-code`, `resend-mfa-code`, `reset-password-2fa-sms` & `reset-password-check-email` pages now include a taxonomy level 2.
- `checkAccountRecoveryPermitted` middleware is applied to `router.post(PATH_NAMES.CHECK_YOUR_EMAIL_CHANGE_SECURITY_CODES, ...);` which sets the `isAccountRecoveryJourney` user session value. This is used by the middleware to correctly state the `account recovery` taxonomy level 2 in the `check-your-email-security-codes` POST page.
- `check-your-email-controller.ts` contained a series of unused `contentId`s and `taxonomyLevel2`s for a set of different journey types. Only the `createAccount` ones are used, so rest are removed.
  - The GET controller also now sets the `isAccountCreationJourney` user session value so it can be used by the middleware.
  - Some pages on `check-your-email` were previously setting taxonomy level 1 to `accounts`, where it is now correctly `authentication`.