AUT-3760: Apply reauth taxonomies
Ensure when a user is reauthenticating that pages include the `re auth` taxonomy. This is so data analytics can determine the context in why a user requested a page.

In addition to the logic changes, integration tests are added on key pages to ensure they log the correct taxonomies for the reauth or sign in contexts.