BAU: Snapshot current contentIds
Adjust the method that we use to test snapshots of taxonomies to also support snapshots of contentIds across our existing integration test suite.

This will allow us an obvious way to see changes have been made on pages as we come to refactor how contentIds are set and add reauthentication contentIds.