AUT-3760: Add utility to get contentId from request
Currently contentIds are defined ad-hoc, in templates or controllers. This utility method is the start of minor change to where content ids are defined. The function looks at the request and uses the path and user session to determine the content Id to serve.

This is added to a middleware for calculating Google Analytics properties and sending them to the template renderer. This starts to centralise the contentId logic, which will make complex changes and additions easier to implement.

The middleware allows the controllers calling `render` to override the contentId calculated by this utility. This ensures it does not break other existing contentId calculations.

The `/enter-mfa` path is tackled first by refactoring how the content ids of the path recorded.